--- src/routes/footer/index.routes.js
@@ -1,5 +1,0 @@
-import { contactUs } from './contact-us.routes.js'
-
-export const footerRoutes = [
-  contactUs
-]
