+++ test/unit/routes/business/business-legal-status-check-routes.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { getBusinessLegalStatusChange } from '../../../../src/routes/business/business-legal-status-change.routes.js'
+import { getBusinessLegalStatusChange } from '../../../../src/routes/business/business-legal-status-change-routes.js'
 
 describe('change business legal status', () => {
   test('should have the correct method and path', () => {
