+++ test/unit/routes/errors/service-unavailable-routes.test.js
@@ -1,5 +1,5 @@
 import { vi, beforeEach, describe, test, expect } from 'vitest'
-import { serviceUnavailable } from '../../../../src/routes/errors/service-unavailable.routes.js'
+import { serviceUnavailable } from '../../../../src/routes/errors/service-unavailable-routes.js'
 
 const mockView = vi.fn()
 const mockH = {
